Move Terminator struct to a seperate file
